Refactor some key modules into a core module
