+++ src/core/math.rs
