+++ src/lib.rs
@@ -3,7 +3,7 @@ use num_cpus;
 
 #[macro_use]
 pub(crate) mod macros;
-pub(crate) mod math;
+pub(crate) mod core;
 pub(crate) mod ray;
 pub(crate) mod img;
 pub(crate) mod space;
